- Change ProblemDocument.Difficulty to enum type and update all related ElasticSearch logic.
- Add topic existence validation on problem creation with new ITopicRepository and TopicErrors.
- Refactor IExecutionService: rename ExecuteCodeAsync to RunCodeAsync, use TestCasesDto, add single test case overload, and improve output comparison.
- Extend IFileService with IFormFile reading and clarify async signatures.
- Dynamically resolve script file path in Helper for better portability.
- Implement "Run Code" feature: add DTOs, command, handler, and validator for custom test case execution.

+++ src/CodeClash.API/Extensions/ElasticSearchSeeder.cs
@@ -41,7 +41,7 @@ public sealed class ElasticSearchSeeder(
         {
             Id = p.Id,
             Name = p.Name,
-            Difficulty = (int)p.Difficulty,
+            Difficulty = p.Difficulty,
             Topics = p.ProblemTopics
                             .Select(pt => pt.TopicId)  // adjust property name
                             .ToList()
